Linting Error

+++ osstats/collector.go
@@ -111,11 +111,11 @@ func (s *StatsCollector) collectStats() {
 }
 
 func formatGB(val uint64) float64 {
-	return float64(float64(val) / (1024 * 1024 * 1024))
+	return float64(val) / (1024 * 1024 * 1024)
 }
 
 func formatMB(val uint64) float64 {
-	return float64(float64(val) / (1024 * 1024))
+	return float64(val) / (1024 * 1024)
 }
 
 func (s *StatsCollector) get() (*osStat, error) {
